api_flask/requirements.txt
Flask
Flask-Cors
Flask-PyMongo
Flask-Bcrypt
PyJWT
python-dotenv
ollama
pymongo[srv]
uuid